Convert DataClassDesigner jest test from enzyme to RTL
